Fixing the issue where I was trying to include a file inside the testing engine on the refinerycms application
