+++ spec/spec_helper.rb
@@ -1,6 +1,31 @@
 require 'rbconfig'
 require 'factory_girl'
-require File.dirname(__FILE__) + '/../testing/spec/support/controller_macros.rb'
+
+module ControllerMacros
+  def login_user
+    before (:each) do
+      @user = Factory(:user)
+      @request.env["devise.mapping"] = Devise.mappings[:admin]
+      sign_in @user
+    end
+  end
+
+  def login_refinery_user
+    before (:each) do
+      @refinery_user = Factory(:refinery_user)
+      @request.env["devise.mapping"] = Devise.mappings[:admin]
+      sign_in @refinery_user
+    end
+  end
+
+  def login_refinery_translator
+    before (:each) do
+      @refinery_translator = Factory(:refinery_translator)
+      @request.env["devise.mapping"] = Devise.mappings[:admin]
+      sign_in @refinery_translator
+    end
+  end
+end
 
 def setup_environment
   # This file is copied to ~/spec when you run 'rails generate rspec'
